+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 ### Added
 
 - Added textures for Living Flames from Dungeons Delight
+
+### Fixed
+
+- Fixed flame particle rotation not working when using Sodium